refactor: move Header out of ondisk to break its cycle with key_spaces
# Summary

`Header` moves to its own top-level module and gains the `Header::KEY` const
that used to be `OnDisk::KEY_HEADER`. With that, `key_spaces` no longer
depends on `ondisk`, and the crate's production module graph is acyclic.

# Details

`key_spaces` needed `ondisk` twice: for `Header`, the value type of the
`DataHeader` key space, and for the key that header is stored under. Both are
properties of the header record rather than of the upgrade machinery in
`OnDisk`, so both now live with the type. This follows the existing shape of
`state` and `state_machine`, which hold sled value types in their own modules
that `key_spaces` then references.

`Header` stays out of `data_version` despite describing a version, because it
is sled-serialized and `data_version` must remain free of the sled dependency
to be usable by configuration code.

This also drops `raft_log_v004`'s dependency on `ondisk`, which existed only
to reach `Header` from a test.

`Header` moves from `databend_meta_raft_store::ondisk` to
`databend_meta_raft_store::header`.

With this the seven cycles that motivated the split are gone: every remaining
edge between production modules points one way, and the module set now
topologically sorts.
