+++ crates/server/raft-store/src/lib.rs
@@ -46,6 +46,7 @@ pub mod applier;
 pub mod config;
 pub mod data_version;
 pub mod db_exporter;
+pub mod header;
 pub mod immutable_compactor;
 pub mod key_spaces;
 pub mod leveled_store;
