+++ crates/server/service/src/store/store.rs
@@ -23,8 +23,8 @@ use anyerror::AnyError;
 use databend_meta_raft_store::MetaStartupError;
 use databend_meta_raft_store::config::RaftConfig;
 use databend_meta_raft_store::db_exporter::DBExporter;
+use databend_meta_raft_store::header::Header;
 use databend_meta_raft_store::key_spaces::RaftStoreEntry;
-use databend_meta_raft_store::ondisk::Header;
 use databend_meta_raft_store::ondisk::TREE_HEADER;
 use databend_meta_raft_store::raft_log_v004;
 use databend_meta_raft_store::raft_log_v004::Cw;
